jlc can now run via middleware or new management command
