+++ .gitignore
@@ -1,5 +1,6 @@
 .DS_Store
 *.py[co]
+*~
 
 # Packages
 *.egg
